chore: 🤖 davidl-fr quick-wins janitor pass (#001)
- fix jest setupFilesAfterEnv key (was setupFilesAfterEach)
- remove unused deps react-helmet and react-mailchimp-subscribe
- fix PWA theme-color to use media-query meta tags + darkTheme.bg.accent
- convert talks.tsx getServerSideProps → getStaticProps with revalidate:3600
- fix 3 stale test assertions (ecosystem-06, expo-router-05, countChallenges=22)
- lazy-load Hero Typewriter via next/dynamic (ssr:false)
- migrate next/legacy/image → next/image across 5 files
- update README: monorepo commands, yarn version, drop stale sections
