+++ apps/davidl-fr/src/pages/api/workshop.test.ts
@@ -3,6 +3,6 @@ import { allWorkshops, countChallenges } from '~/pages/api/workshop';
 describe('workshop', () => {
   it('should count the number of challengers to display "Challenges completed"', () => {
     const numberChallenges = countChallenges(allWorkshops);
-    expect(numberChallenges).toBe(21);
+    expect(numberChallenges).toBe(22);
   });
 });
