+++ package.json
@@ -85,10 +85,8 @@
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-gravatar": "^2.6.3",
-    "react-helmet": "^6.1.0",
     "react-intersection-observer": "^9.16.0",
     "react-lite-youtube-embed": "^2.4.0",
-    "react-mailchimp-subscribe": "^2.1.3",
     "react-markdown": "^10.0.0",
     "react-typed": "^1.2.0",
     "regenerator-runtime": "0.14.1",
@@ -139,7 +137,6 @@
     "@types/react": "19.0.10",
     "@types/react-dom": "19.0.4",
     "@types/react-gravatar": "^2.6.14",
-    "@types/react-helmet": "^6.1.11",
     "@types/styled-components": "^5.1.34",
     "@typescript-eslint/eslint-plugin": "8.46.2",
     "@typescript-eslint/parser": "8.46.2",
